src/requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
Flask==2.2.2
flask-restx==1.0.6
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pydantic==1.10.4
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2022.7.1
SQLAlchemy==2.0.3
typing_extensions==4.4.0
Werkzeug==2.2.2
Flask-SQLAlchemy==3.0.3
redis==4.0.0
marshmallow==3.19.0
pytest==7.2.1
pytest-asyncio==0.12.0
aiohttp==3.7.2

